c_begin_list(): The [] are not always necessary

+++ Source/env.c
@@ -70,6 +70,7 @@
 *    fd  Mar 05: c_begin_document() debugged for titdat content
 *    ggs Apr 04: c_begin_list() delete compressed and no_compressed
 *    ggs Apr 12: c_begin_list() test the compressed and not only short
+*    ggs Apr 21: c_begin_list(): The [] are not always necessary
 *
 ******************************************|************************************/
 
@@ -2562,7 +2563,8 @@ int       listkind)     /* List type */
          /* Siehste, hat's mal wieder an den Anfang gesetzt      */
          /* Das mitzuentfernende Leerzeichen kommt von tokcpy2() */
          
-         delete_once(sWidth, "!short ");
+         if ( !delete_once(sWidth, "!short ") )
+            delete_once(sWidth, "!short");
       }
       else
       {
@@ -2586,7 +2588,8 @@ int       listkind)     /* List type */
             /* Siehste, hat's mal wieder an den Anfang gesetzt      */
             /* Das mitzuentfernende Leerzeichen kommt von tokcpy2() */
          
-            delete_once(sWidth, "!compressed ");
+            if ( !delete_once(sWidth, "!compressed ") )
+               delete_once(sWidth, "!compressed");
          }
          else
          {
@@ -2594,6 +2597,32 @@ int       listkind)     /* List type */
             delete_last(sWidth, " !compressed");
          }
       }
+     else
+     {
+        ptr = strstr(sWidth, "!not_compressed");
+        
+        if (ptr != NULL)
+        {
+           /* Aha, !not_compressed wird benutzt. Da manche Dumpfnasen das aber nicht */
+           /* immer ans Ende setzen, hier gleich die passenden Abfragen.         */
+
+           strcpy(sShort, "!not_compressed");   /* Fuer set_env_compressed() */
+
+           if (ptr == sWidth)
+           {
+              /* Siehste, hat's mal wieder an den Anfang gesetzt      */
+              /* Das mitzuentfernende Leerzeichen kommt von tokcpy2() */
+         
+              if ( !delete_once(sWidth, "!not_compressed ") )
+                 delete_once(sWidth, "!not_compressed");
+           }
+           else
+           {
+              /* So gehoert es sich, schoen ans Ende der Zeile */
+              delete_last(sWidth, " !not_compressed");
+           }
+        }
+     }
    }
    
    del_internal_styles(sWidth);
@@ -2601,9 +2630,6 @@ int       listkind)     /* List type */
    replace_udo_quotes(sWidth);
 
    set_env_compressed(iEnvLevel, sShort); /* check the compressed flag */
-   ptr = strstr(sWidth, "!not_compressed");
-   if ( ptr != NULL )
-      delete_last(sWidth, " !not_compressed");
 
    switch (desttype)
    {
